EN-870: values in ESC status were moved to a new Envgo fact group.
