+++ qgroundcontrol.pro
@@ -434,6 +434,7 @@ HEADERS += \
     src/QmlControls/QmlUnitsConversion.h \
     src/Vehicle/VehicleEscStatusFactGroup.h \
     src/Vehicle/VehicleEscInfoFactGroup.h \
+    src/Vehicle/VehicleEnvgoFactGroup.h \
     src/api/QGCCorePlugin.h \
     src/api/QGCOptions.h \
     src/api/QGCSettings.h \
@@ -448,6 +449,7 @@ contains (DEFINES, QGC_ENABLE_PAIRING) {
 SOURCES += \
     src/Vehicle/VehicleEscStatusFactGroup.cc \
     src/Vehicle/VehicleEscInfoFactGroup.cc \
+    src/Vehicle/VehicleEnvgoFactGroup.cc \
     src/api/QGCCorePlugin.cc \
     src/api/QGCOptions.cc \
     src/api/QGCSettings.cc \
